Seperate quotes into cog component
